ui/src/bigtrace/requirements.txt
flask
google-cloud-secret-manager
httpx
google-auth
gunicorn
pyjwt
cryptography